Adjusted the logic for when a mob is drawn above another mob.
- Basically, held mobs are always drawn in a way that respects their Z, and "above_holder" is just a pseudo-hack to force them being drawn above.

+++ source/source/game_states/gameplay/pause_menu.cpp
@@ -2459,7 +2459,7 @@ void pause_menu_t::init_status_page() {
             game.states.gameplay->get_amount_of_field_pikmin(pt_ptr);
         long in_onion =
             game.states.gameplay->get_amount_of_onion_pikmin(pt_ptr);
-        long total = on_field + in_onion;
+        long total = (long) on_field + in_onion;
         
         long new_piks = 0;
         auto new_it =
